updates: Change categories to labels
Mengubah nama fitur kategori menjadi label, karena sebelumnya sudah terdapat kategori pada tabel budget
(kebutuhan, keinginan, dan tabungan)

+++ database/migrations/2026_03_14_175414_create_budgets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->decimal('total_income', 12, 2);
+            $table->integer('total_income');
 
             $table->tinyInteger('month');
             $table->year('year');
